+++ ChangeLog.md
@@ -1,9 +1,13 @@
 # BountyBot Change Log
 ## BountyBot Version 2.11.0bfi:
-- Fixed Goal Point contributions not showing up in reward messages
+### Reaction Toasts
+- Bounty Hunters can now raise a messageless toast by reacting to another Bounty Hunter's discord message with 🥂
+- Reaction toasts can be toggled off in `/server-config`
 ### Premium Features
 - Added premium commands `/festival start-gp` and `/festival close-gp` for toggling times where gp contributions are multiplied (old festival commands renamed to `/festival start-xp` and `/festival close-xp`)
 - Added the `nickname` option to `/config-premium` to allow bot managers to set a nickname for BountyBot that won't get cleared by toggling festivals on or off
+### Bug Fixes
+- Fixed Goal Point contributions not showing up in reward messages
 ## BountyBot Version 2.10.1bi:
 This version includes the addition of the "b" version category, for "bug fixes".
 - Fixed several crashes related to missing permissions, modal timeouts, trying to DM users who've blocked BountyBot, and residual data from former server members